+++ ChangeLog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- (Export) Occlusion map support
 - (Documentation) XML documentation comments on many types
 - (Documentation) Initial setup for DocFX generator
 ### Changed